Consolidate AssetTicket business logic into asset-tickets.ts
Moves cascadeDoneStatus and the mitigation router's inline device-group
resolution loop into asset-tickets.ts alongside createAssetTicket,
which now also accepts an optional externalMapping so Fleet's per-asset
loop is one write instead of a create-then-update. detachAsset reuses
the asset it already fetched instead of re-querying it for the
activity label.
